requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==4.2.18
django-allauth==65.3.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.2
pillow==10.2.0
propcache==0.2.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
stripe==7.11.0
twilio==9.4.3
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.6.0
yarl==1.15.2
